Refactoring: GridCreatingConditionTriangle::MovePolyLineVertexCommand implementation is separated from gridcreatingconditiontriangle.cpp.
