+++ libs/gridcreatingcondition/triangle/triangle.pro
@@ -114,7 +114,8 @@ HEADERS += gcc_triangle_global.h \
            private/gridcreatingconditiontriangle_finishpolylinedefiningcommand.h \
            private/gridcreatingconditiontriangle_movepolygoncommand.h \
            private/gridcreatingconditiontriangle_movepolygonvertexcommand.h \
-           private/gridcreatingconditiontriangle_movepolylinecommand.h
+           private/gridcreatingconditiontriangle_movepolylinecommand.h \
+           private/gridcreatingconditiontriangle_movepolylinevertexcommand.h
 FORMS += gridcreatingconditiontrianglepolygoncoordinateseditdialog.ui \
          gridcreatingconditiontrianglepolylinecoordinateseditdialog.ui \
          gridcreatingconditiontriangleremeshdialog.ui \
@@ -139,7 +140,8 @@ SOURCES += gridcreatingconditioncreatortriangle.cpp \
            private/gridcreatingconditiontriangle_finishpolylinedefiningcommand.cpp \
            private/gridcreatingconditiontriangle_movepolygoncommand.cpp \
            private/gridcreatingconditiontriangle_movepolygonvertexcommand.cpp \
-           private/gridcreatingconditiontriangle_movepolylinecommand.cpp
+           private/gridcreatingconditiontriangle_movepolylinecommand.cpp \
+           private/gridcreatingconditiontriangle_movepolylinevertexcommand.cpp
 TRANSLATIONS += languages/iricGccTriangle_es_ES.ts \
                 languages/iricGccTriangle_fr_FR.ts \
                 languages/iricGccTriangle_id_ID.ts \
